requirements.txt
Flask==2.3.2
flasgger>=0.9.5
apispec==6.3.0
Flask-marshmallow>=0.14.0
marshmallow==3.20.1
pathlib~=1.0.1
uuid~=1.30
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
gunicorn==21.2.0
gunicorn_color==0.1.0